Use a Kotlin sealed class for BackendName
This is a toe in the water towards moving more of the app towards
Kotlin longer-term.

The immediate advantage is it separates the members of BackendName from
the generated list itself (BackendName.kt vs BackendNames.kt), so there
are no longer non-trivial source lines in a big string inside
build.gradle.kts - while still maintaining the enum-like aspect. I
don't think that split across files can be achieved in a Java enum.