Enable Compose, minSdk 21 (=5.0), targetSdk 34
I'll be using Compose for the new keyboard view and then rolling it out
to other areas of the app.

According to apilevels.com this takes the proportion of devices supported
from 99.64% down to 99.34% (and increasing all the time of course).
Or Android Studio's "Help me choose" window puts 5.0 at 99.5%.