Support predictive back gesture; remove right-mouse-button tweak
Android previously sent KEYCODE_BACK events from the right mouse
button, and in 2016 we accepted a kind patch to fix this. It seems
as of Android 11 (2020) Android no longer does this, so on 11+,
don't attempt that handling, and hide the preference. This avoids
warnings that we're attempting handling that doesn't work with
predictive back, and sets us up to remove these methods completely
when minSdk reaches 30.