Stop rapidly en/disabling onBackPressedCallback!
On key press, we disable Back and sendMessageDelayed to re-enable it
after 600ms to prevent accidentally backing out of the game. If
there's another key press during this time, we need to drop the
previous delayed message as well as sending a fresh one!  This was
especially bad during drags, because those are a continuous stream of
key presses.

Fixes #631