Clarify documentation of draw_polygon()'s coordinate format.
Previously, this function's documentation just stated that `coords`
contained the X and Y coordinates of the polygon's vertices, without
specifying in what order they were listed. Thus, it would have been
reasonable to (wrongly) assume that all the X coordinates were listed
in the first half of the array, followed by all the Y coordinates.

Now it's clear that each point's X and Y coordinates are stored
directly next to each other, and in that order.