Fix reference to Git commit hash in puzzles.h.
I had tried to make this comment refer to the commit in which the
change in question (the new drawing API semantics) was introduced, but
something about Simon applying the patch caused its hash to change.

Now that the commit in question is in Simon's branch, it should remain
constant forever.