Cosmetic fix for mouse-to-tile coordinate transformation.
Clicking just outside the left or top edges of the grid in Pattern
has the same effect as clicking just inside the same edge.
This happens because the FROMCOORD macro uses integer division,
which rounds towards zero rather than towards negative infinity.

One way to fix this is to add TILE_SIZE before the division
and subtract 1 afterwards.