slant: add preference: fading grounded components
Commit message added by SGT:

The general idea is that the puzzle constraint "you must not form a
loop" can also be phrased as "every grid vertex must have a path to
the boundary of the grid", on the basis that in a grid with every
square filled, any connected component of vertices _not_ joined to the
boundary by a path must instead be surrounded by a loop blocking every
way it could get there. So by changing the colour of each
"grounded" (connected to the boundary) grid edge, you draw the
player's attention to the components that don't yet have a path to the
edge, so they can consider the possible paths.

This is the kind of hint that users can very easily find to be
patronising and intrusive, not to mention a spoiler if they haven't
even made the leap from "no loops" to "everything must be grounded".
So it's a preference, and off by default.

This patch is somewhat rewritten from the submitted version, to
conform to local style and also compile in older C versions.