Describe front-end pixel scaling in devel.but
This adds two paragraphs.  The first describes the model that the GTK
front-end uses, where a "pixel" in the drawing API can map to multiple
pixels in the pixmap used for drawing.  To preserve pixel-accuracy,
this requires that there be an integer number of physical pixels to
each logical pixel, which happily is the case in GTK (at least up to
GTK 3).

The second paragraph attempts to explain how this relates to the
device_pixel_ratio provided to midend_size().  That's used by the
JavaScript front end and can cope with the arbitrary non-integer scale
factors thatr might be provided by Web browsers.  However, it works by
just exposing physical pixels to the back end, which means that thin
lines tend to become hard to see as the physical pixels get small.

I think the best approach is actually a mixture of the two, where an
integer is factored out of the scale factor by the front end, and the
remaining factor (which should be < 2) is passed to midend_size().  I
haven't tried implementing this yet, though, so I might change my
mind.