js: handle most of the device pixel ratio in the front end
This makes the JavaScript front end work more like the GTK front end in
its handling of varying device pixel ratios, which GTK calls scale
factors.  This is also an implementation of what I just described in
devel.but.

If the device pixel ratio is an integer, then the mid and back ends just
ge told about CSS logical pixels and the front end handles all the
necessary scaling itself, much like GTK.  If the device pixel ratio
isn't an integer then this isn't possible.  Instead we handle scaling by
the floor of the device pixel ratio in the front end, and pass the
remaining scaling factor to the mid end as the device pixel ratio.  This
means that pressing Ctrl-+ in a browser still causes the canvas to grow
nicely, but on reaching a zoom of 200% the thin lines all get thicker.

I think this is a more or less ideal way to handle the device pixel
ratio, and in particular produces decent results without any effort in
back ends and without complicating the drawing API.  It does need rather
a lot of front-end work, though.