js: trust CanvasRenderingContext2D.drawImage() to DTRT
As far as I can tell, the specification for
CanvasRenderingContext2D.drawImage() requires it to sensibly handle
cases where the rectangle it's working on isn't entirely contained in
both the source and destination canvases.  So there's not really any
reason for Puzzles to carefully trim the region being copied to their
intersection.  This vastly simplifies a lot of the blitter-handling code
because it can just ask the user agent to copy the whole blitter back
and forth.

I've checked archived versions of the HTML spec back to 2014, and they
all say this should work.  MDN doesn't suggest otherwise either.  So I
think if this code was ever necessary, it's not any more.