GTK: avoid explicitly using Cairo image surfaces
Where we want to create a surface that's suitable for copying to and
from another surface, it's better to use cairo_surface_create_similar()
based on our window rather than explicitly creating an image surface.
This has two major advantages.  First, it means that the surface is one
that can be copied easily to the target window (usually an Xlib surface
when running under X, for instance).  Second, the new surface naturally
inherits the device scale from the original.

The first of these means that there's no longer any benefit in having a
separate pixmap when using Cairo, so USE_CAIRO_WITHOUT_PIXMAP becomes
equivalent to USE_CAIRO.  The second means that a lot of code for
explicitly handling scale factors can be removed.  Essentially the only
remaining reason for recording the scale factor is to detect when it
changes and re-create the backing surface.

There is one place where we still create an image surface, and that's
when running in headless mode to create icons.