Remove BLITTER_FROMSAVED from drawing API
blitter_load had a feature whereby passing in BLITTER_FROMSAVED as both
x and y co-ordinates would cause them to be treated as whatever was last
passed to blitter_save() for the same blitter.  But no back end has ever
used this feature, and the JavaScript front end has never implemented
it, so it's obviously not a very memorable feature.  Moreover, it would
be quite reasonable to want to load from a blitter at (-1, -1), so
usurping those co-ordinates could cause surprising results.

I think if one did want this feature, it would be better implemented as
a separate function, but it doesn't seem like much trouble to let the
back end remember the co-ordinates.

I don't think this requires a change to the drawing API version number
because it will cause a compile failure in the (unlikely) event that
any third-party front end is actually using the feature.