GTK: correct offset when copying backing to window
When I removed the explicit code for handling non-unity scale factors
in favour of letting Cairo do the hard work, I accidentally simplified
draw_area() a little too much and removed the offsets that allow for
the fact that the puzzle surface doesn't necessarily fill the window,
meaning that the wrong parts of the surface tended to get painted into
the window.  I've now corrected this.