Mines: put a warning on the web page about bug reports.
Mines is the puzzle that has attracted by far the largest number of
bug reports, including a great many claims that the "no guessing"
feature wasn't living up to its claims. But a bug has never actually
been found in the grid generation code. (The usual number of other
kinds of bug, just not _that_ bug.)

I've wondered for a while about putting some prominent text on the web
version of the puzzle stressing just how many of that kind of bug
report I've received and that all of them have been wrong, just to
encourage people to _try_ thinking again. So I went through my email
records and counted them. I make it 59 as of today, which is a slight
surprise, because it _felt_ more like 100-200 to me. Still, "over 50"
is a reasonable thing to write in the

It's an interesting question _why_ Mines gets all these false-positive
reports. One obvious possibility is simply that it's more popular than
the rest of the games in this collection. But a more interesting idea
is that perhaps players of ordinary Minesweeper, striving to get a
fast clearance time, are accustomed to not spending a lot of time
thinking about problems - instead, either open something quickly or
guess quickly, because if you guess you _might_ not get a good
time (kaboom!), but if you ponder for 30 seconds you _definitely_
won't get a good time.