Palisade: preference to clear complete regions.
Sometimes you complete a Palisade region by ruling out all its
internal edges one by one, clicking to change the initial "don't know"
line state into "definitely not a region border", and then you fill in
the edges. But sometimes a region is completed for another reason,
e.g. you completed all its surrounding regions before you'd thought
about this one particularly. Then it looks ugly to have the "don't
know" edges in a region that's finished already.

So here's a preference option to make completed regions look nicer. In
a complete region of exactly the right size, the internal edges are
shown as "definitely not" even if their logical state is "don't know".
It's very similar to Light Up's preference to stop showing the manual
"not a light" marks once a square is lit up: it means a fully solved
grid always looks nice, whether you bothered to clean it up by hand or
not.

This is a bit weird because it's a _display_-time change, so clicking
in the region won't change edge states back and forth - or rather, it
will, but you won't see the difference. Also it's slightly
patronising. So it's a user preference, and off by default.